requirements.txt
# Core dependencies for Quickscene video timestamp-retrieval system

# Audio Speech Recognition
openai-whisper==20250625
assemblyai==0.17.0

# Text Embedding and NLP
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.5.1

# Vector Search and Indexing
faiss-cpu==1.9.0

# Data Processing
numpy==2.2.1
pandas==2.2.3

# Configuration Management
PyYAML==6.0.2

# Development and Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-benchmark==5.1.0

# Logging and Monitoring
loguru==0.7.3

# Optional: Enhanced audio processing
librosa==0.10.2
soundfile==0.12.1

# Optional: Progress bars and CLI
tqdm==4.67.1
click==8.1.8

# Web API Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20
jinja2==3.1.4
